+++ VSProjTypeExtractor/VSProjTypeExtractor.cpp
@@ -1,3 +1,32 @@
+/*
+    VSProjTypeExtractor - Visual Studio project type GUID extractor
+    VSProjTypeExtractor.cpp - Implementation File
+    Copyright (c) 2019, Lucian Muresan.
+
+    MIT License
+
+    Permission is hereby granted, free of charge, to any person obtaining a copy
+    of this software and associated documentation files (the "Software"), to deal
+    in the Software without restriction, including without limitation the rights
+    to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+    copies of the Software, and to permit persons to whom the Software is
+    furnished to do so, subject to the following conditions:
+
+    The above copyright notice and this permission notice shall be included in all
+    copies or substantial portions of the Software.
+
+    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+    SOFTWARE.
+
+    You can contact the author at :
+    - VSProjTypeExtractor homepage and source repository : https://github.com/lucianm/VSProjTypeExtractor
+*/
+
 #include "VSProjTypeExtractor.h"
 
 #include <msclr/marshal.h>
@@ -16,10 +45,10 @@ namespace VSProjTypeExtractor {
 		VSProjTypeExtractorManaged::VSProjTypeWorker m_managedWorker;
 	public:
 		static property ClassWorker^ Instance { ClassWorker^ get() { return % m_instance; } }
-		System::String^ GetProjTypeGuidStringManaged(System::String^ projPath, unsigned int vs_MajVer )
+		System::String^ GetProjTypeGuidStringManaged(System::String^ projPath)
 		{
 			msclr::lock lock(this);
-			return m_managedWorker.ExtractProjectTypeGuid(projPath, vs_MajVer);
+			return m_managedWorker.ExtractProjectTypeGuid(projPath);
 		}
 		void CleanUp()
 		{
@@ -30,12 +59,12 @@ namespace VSProjTypeExtractor {
 }
 
 
-bool Vspte_GetProjTypeGuidString(const char* projPath, char* projTypeGuid, unsigned int projTypeGuidMaxLength, unsigned int VS_MajorVersion)
+bool Vspte_GetProjTypeGuidString(const char* projPath, char* projTypeGuid, unsigned int projTypeGuidMaxLength)
 {
 	bool bSuccess = false;
 	System::String^ strProjPath = gcnew System::String(projPath);
 	try {
-		System::String^ strProjTypeGuid = VSProjTypeExtractor::ClassWorker::Instance->GetProjTypeGuidStringManaged(strProjPath, VS_MajorVersion);
+		System::String^ strProjTypeGuid = VSProjTypeExtractor::ClassWorker::Instance->GetProjTypeGuidStringManaged(strProjPath);
 
 		if (strProjTypeGuid->Length >= VSPROJ_TYPEEXTRACT_MAXGUID_LENGTH - 1 && unsigned int(strProjTypeGuid->Length) <= projTypeGuidMaxLength)
 		{
@@ -48,7 +77,7 @@ bool Vspte_GetProjTypeGuidString(const char* projPath, char* projTypeGuid, unsig
 	}
 	catch (System::Exception^ e)
 	{
-		System::Console::WriteLine("\n{0}\noccured for project file '{1}' loaded in Visual Studio {2}", e->ToString(), strProjPath, VS_MajorVersion);
+		System::Console::WriteLine("\n{0}\noccured for project file '{1}' loaded in Visual Studio {2}", e->ToString(), strProjPath);
 		if (e->InnerException)
 		{
 			System::Console::WriteLine("\nInner exception: {0}\n", e->InnerException->ToString());